Apply suggestions from code review

+++ include/amici/model.h
@@ -648,7 +648,7 @@ class Model : public AbstractModel, public ModelDimensions {
      * activated, this lower boundary must ensure that log(sigma) + min_sigma > 0.
      * @param min_sigma lower boundary
      */
-    void setMinimumSigmaResiduals(float min_sigma) {
+    void setMinimumSigmaResiduals(double min_sigma) {
         min_sigma_ = min_sigma;
     }
     
@@ -662,7 +662,8 @@ class Model : public AbstractModel, public ModelDimensions {
     
     /**
      * @brief Specifies whether residuals should be added to account for parameter dependent sigma.
-     * If set to true, additional residuals of the form `sqrt{log(\sigma) + C}` will be added. This enables
+     *
+     * If set to true, additional residuals of the form \f[ \sqrt{\log(\sigma) + C} \f] will be added. This enables
      * least-squares optimization for variables with Gaussian noise assumption and parameter dependent
      * standard deviation sigma. The constant C can be set via :meth:`setMinimumSigmaResiduals`.
      * @param sigma_res if true, additional residuals are added
